backend/requirements.txt
auto_mix_prep==0.2.0
fastapi==0.123.5
langchain==1.1.0
langchain_community==0.4.1
langchain_mongodb==0.8.0
langchain_openai==1.1.0
langchain_text_splitters==1.0.0
pydantic==2.12.5
pydantic_settings==2.12.0
pymongo==4.15.5
python-dotenv==1.2.1
uvicorn==0.38.0
